sorting a sentence
Okay so the concept is to first split the string to a list of word and now each word has a number at its end and that will be our sorting order in order to tell python to sort the list based on the number in the end we do this list1 = sorted(list1, key=lambda x:int(x[-1])) so we are trying to say that sort the list using the int value of the last char for each element x in the list1, that;s the explanation for the python lambda function. Once the list is sorted we clip the number part for each element in the list and return the string it will form using " ".join(list1).